+++ moqt/internal/message/group.go
@@ -24,7 +24,7 @@ func (g GroupMessage) Encode(w io.Writer) error {
 	b := pool.Get(msgLen + VarintLen(uint64(msgLen)))
 	defer pool.Put(b)
 
-	b, _ = WriteMessageLength(b, uint16(msgLen))
+	b, _ = WriteMessageLength(b, uint64(msgLen))
 	b, _ = WriteVarint(b, g.SubscribeID)
 	b, _ = WriteVarint(b, g.GroupSequence)
