+++ moqt/internal/message/session_server.go
@@ -33,7 +33,7 @@ func (ssm SessionServerMessage) Encode(w io.Writer) error {
 	b := pool.Get(msgLen + VarintLen(uint64(msgLen)))
 	defer pool.Put(b)
 
-	b, _ = WriteMessageLength(b, uint16(msgLen))
+	b, _ = WriteMessageLength(b, uint64(msgLen))
 	b, _ = WriteVarint(b, uint64(ssm.SelectedVersion))
 
 	// Append parameters
