+++ moqt/internal/message/subscribe.go
@@ -35,7 +35,7 @@ func (s SubscribeMessage) Encode(w io.Writer) error {
 	b := pool.Get(msgLen + VarintLen(uint64(msgLen)))
 	defer pool.Put(b)
 
-	b, _ = WriteMessageLength(b, uint16(msgLen))
+	b, _ = WriteMessageLength(b, uint64(msgLen))
 	b, _ = WriteVarint(b, uint64(s.SubscribeID))
 	b, _ = WriteVarint(b, uint64(len(s.BroadcastPath)))
 	b = append(b, s.BroadcastPath...)
